distgraphconvert support for 4294967296 nodes max

+++ tools/dist-graph-convert/dist-graph-convert-helpers.h
@@ -136,8 +136,8 @@ std::vector<uint32_t> loadEdgesFromEdgeList(std::ifstream& edgeListFile,
   // read lines until last byte
   edgeListFile.seekg(localStartByte);
   while ((uint64_t)(edgeListFile.tellg() + 1) != localEndByte) {
-    uint32_t src;
-    uint32_t dst;
+    uint64_t src;
+    uint64_t dst;
     edgeListFile >> src >> dst;
     if (startAtOne) {
       src--;
